Update README.md
An evaluation of the company's sales performance in 2023 is underway, with a focus on various elements such as segments, states, codes, category and products. The primary goal is to identify areas necessitating heightened attention. Recommendations will be formulated to guide the company in enhancing performance and making informed decisions for progress. The analysis seeks to yield KPIs based on the company's data, that will play a central role in driving innovation and addressing inquiries that will facilitate a more productive year in the future.

Key Questions That Need Answers

Quantity of units sold for each product category
States with the highest number of transaction
Total Profits made from each region	
Total revenue by each region
Variation and highest ship mode utilized for products
Subcategory with the highest sales
Quantity of goods requested in each category
Discount given to each category